% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibo.R
\name{fibo}
\alias{fibo}
\title{This calculates the nth fibonacci number}
\usage{
fibo(n)
}
\arguments{
\item{n}{positive integer indicating the fibonacci number you want to compute}
}
\value{
integer, the nth fibonacci number
}
\description{
This calculates the nth fibonacci number
}
\examples{
fibo(1)  # 1
fibo(3)  # 2
}
